refactor(ssh): decompose SSHTerminalView per Omega layout; trim chrome to endpoint strip
Three structural moves aligning with the omega-swift-agent skill's
conventions:

1. Extract TerminalBellController (an @MainActor ObservableObject)
   and the fireAudibleBell() free function from SSHTerminalView.swift
   into Pulse/Views/Terminal/TerminalBellController.swift.
   One-type-per-file rule for non-view types; the controller's
   lifecycle is independent and referenced from one place (the
   view's @StateObject).

2. Move the inline `.toolbar { ... }` block on body into a named
   `@ToolbarContentBuilder private var sshTerminalToolbar`. Body
   shrinks to a clean composition; the three explicit
   ToolbarItem(id:) declarations are grouped under one named seam.

3. Replace the in-view `statusBar` with `connectionEndpointStrip` —
   a single-line monospaced caption showing `host:port` at
   `.connected` only, empty elsewhere. Closes the duplication the
   prior status bar introduced (device name in both toolbar and
   in-view header; "Connected" pill duplicating "Connected to
   <host>" prefix) while preserving the operator-visible host
   string. The host string is the "where am I typing" check that
   closes two session-pinning threats the device-name-alone
   surface cannot: mistaken-target keystrokes under naming
   collisions (template-deployed NetBox sites produce visually
   similar device names), and mid-session re-IP scenarios where
   NetBox mutates Device.primaryIP while a session is connected.

What stayed in place:

- runConnectionLifecycle. The @State capture footprint is too
  wide for clean extraction without a wholesale move to an
  SSHTerminalConnectionViewModel: ObservableObject — flagged as a
  future slice.
- ConnectionStatus, ConnectionAttempt, Connection nested enums.
  Tightly typed view-contract surfaces; breaking them out would
  force test imports to qualify the names.
- The status pill copy contract via Self.statusPillCopy(for:) and
  the primary-action contract via Self.primaryActionShape(for:);
  both extracted as static funcs in the prior commit with full
  ConnectionStatus coverage tests.

ADR Slice 8c note appended covering: click-to-update root cause
and chosen remediation; the connectionEndpointStrip security
rationale (session-pinning under naming collisions and
mid-session re-IP); the decomposition layout; the new test seams;
and the reconnect-button-bypasses-form behaviour as a flagged
edge case.

macOS Debug + iOS Simulator Debug both BUILD SUCCEEDED.
Structural gates clean (TerminalBellController moved out of the
view file; new file exists; statusBar removed; connectionEndpointStrip
present; sshTerminalToolbar present; .toolbar applied exactly once
on body; both static helpers present).
