+++ docs/architecture/0001-ssh-terminal-and-web-foundations.md
@@ -419,6 +419,30 @@ Both must return empty (modulo doc-comment text).
 
 **Slice 8b: nominal-struct wrapper for window routing.** The right structural follow-up: wrap `Device.ID` and `Site.ID` in distinct nominal struct types (`DeviceWindowTarget { let id: Device.ID }`, `SiteWindowTarget { let id: Site.ID }`) so the WindowGroup value type is unambiguous and a future `WindowGroup(for: Int64.self)` addition becomes a compile error rather than a silent mis-route. Cost: ~5 call-site updates, 2 small struct types, one-time state-restoration reset (the WindowGroup value type changes; SwiftUI's restoration storage from the prior build no longer decodes). Benefit: collision becomes a compile error rather than a discipline gate. Deferred to a follow-up slice; the symmetric `id:` fix in Slice 8a is sufficient for today's bug and for ordering changes in `PulseApp.body`.
 
+## Slice 8c note — chrome trim, decomposition, click-to-update fix
+
+Slice 8c lands three changes on top of the Slice 8a window:
+
+**Click-to-update root cause.** Slice 8a's `.windowToolbarStyle(.unifiedCompact(showsTitle: true))` reintroduced a regression originally fixed by commit `cbf1797` (which switched `PulseTerminalAdapter.drainPendingFeed` from SwiftTerm's engine-level `Terminal.feed(buffer:)` to the view-level `TerminalView.feed(byteArray:)` so `feedFinish() → queuePendingDisplay()` schedules an AppKit redraw cycle). Symptom: server output stops repainting until the operator clicks in the window. Diagnosis: `PulseTerminalAdapter.swift:192` still has the load-bearing view-level call; both production and debug entry paths instantiate the same `SSHTerminalView` with identical lifecycle wiring; the debug scene carries no `.windowToolbarStyle` modifier and does not exhibit the symptom. The unified-compact style interferes with `setNeedsDisplay` propagation — likely a responder-chain or layer-tree interaction specific to that style. **Remediation**: switch to `.windowToolbarStyle(.unified(showsTitle: true))`, which preserves the inline-title chrome at slightly taller height without the rendering interference. **Diagnostic tooling**: three `#if DEBUG` `Logger` calls now sit at the render-path boundary on category `ssh.render` — `drainPendingFeed` (byte count + view-attachment state), `makeNSView`/`makeUIView` (fresh TerminalView construction), and `updateNSView`/`updateUIView` (reconciliation). Capture with `log stream --predicate 'subsystem == "pulse" AND category == "ssh.render"'`. Three signatures pick a future fix shape: drained-without-update (display-scheduling regression), repeated makeNSView (identity churn from `body` re-evaluation), or no-drain (MainActor hop deferred). Release builds compile the loggers out entirely.
+
+**Chrome trim — `connectionEndpointStrip` replaces in-view `statusBar`.** Slice 8a's in-view status bar duplicated information surfaced by the toolbar (device name in the navigationTitle, status word in the status pill). Naive removal would have lost the operator-visible host string entirely — the toolbar promotes name and status word but not the resolved host. This slice **trims** rather than removes: a single-line monospaced caption `connectionEndpointStrip` shows `host:port` at `.connected` only, empty elsewhere. Closes two session-pinning threats the device-name-alone surface cannot:
+
+- **Mistaken-target keystrokes.** Two terminals open to similarly-named devices (template-deployed NetBox sites produce names like `OMG-0001-01-SR01` and `OMG-0002-01-SR01`); operator Cmd-Tabs and types a destructive command into the wrong shell. The host string is the structural "where am I typing" check; the device name alone is not sufficient under naming-collision conditions.
+- **Mid-session re-IP.** A NetBox sync mutates `Device.primaryIP` while the session is connected; the endpoint strip shows the actual connected host (resolved at connect time), giving the operator visible evidence the row and the session have diverged.
+
+The accessibility label spells out "Connected endpoint: \(host) port \(port)" so VoiceOver readers get the same signal.
+
+**Decomposition per Omega Swift conventions.** Two structural moves:
+
+- `TerminalBellController` (an `@MainActor ObservableObject`) and the `fireAudibleBell()` free function moved out of `SSHTerminalView.swift` into a new `Pulse/Views/Terminal/TerminalBellController.swift`. One-type-per-file per the `omega-swift-agent` skill's guidance; the controller's lifecycle is independent and it's referenced from one place (the view's `@StateObject`).
+- The inline `.toolbar { ... }` block on `body` moved into a named `@ToolbarContentBuilder private var sshTerminalToolbar: some ToolbarContent`. `body` shrinks to a clean composition; the toolbar's three explicit `ToolbarItem(id:)` declarations are now grouped under one named seam rather than scattered across the body's modifier chain.
+
+`runConnectionLifecycle` and the `ConnectionStatus` / `ConnectionAttempt` / `Connection` nested types stayed in place: the lifecycle's `@State` capture footprint is too wide for clean extraction, and the nested enums are tightly typed view-contract surfaces that gain nothing from breaking out. A future slice may introduce an `SSHTerminalConnectionViewModel: ObservableObject` to own the lifecycle and unlock both extraction and independent testability.
+
+**New test seams.** `SSHTerminalView.statusPillCopy(for: ConnectionStatus) -> String` and `SSHTerminalView.primaryActionShape(for: ConnectionStatus) -> PrimaryActionShape?` extracted as `static func`s. Both are pure mappings over the connection status; both are pinned by unit tests in `SSHConnectFormTests` (under the new "Toolbar contract" MARK section) covering all five `ConnectionStatus` cases. Matches the test-seam pattern established by `autoFireAttempt` (Slice 7) and `applyDeviceDefaultsIfRequested` / `clearDeviceDefaults` (Slice 8). `PrimaryActionShape` is a new local enum (`.disconnect`, `.reconnect`) consumed by `primaryActionButton`'s switch.
+
+**Reconnect-button form re-render — flagged for follow-up.** The `Reconnect` toolbar button calls `submitConnectionAttempt()` directly, re-using the captured form values from the prior attempt. If the operator added a credential during the disconnect interval, the picker would have surfaced it only on a form re-render — but Reconnect bypasses the form. Probably correct (operator memory holds the previous selection), but worth a follow-up if operators report confusion.
+
 ## Forward-looking implementation discipline
 
 The SSH foundations implementation surfaced three structural cascades that were, in retrospect, knowable from public-API inspection before the plan was locked. The rolling pause-decide-resume discipline contained each cascade to one extra commit instead of multiple review rounds, but the cleaner cost reduction is earlier verification.
